Use the right amount of credit4

+++ sandbox/dixond/getpaid.creditpublish/trunk/getpaid/creditpublish/portlets/creditpurchasing/portlet.py
@@ -87,7 +87,7 @@ class Renderer(base.Renderer, RequestMixin):
         buyable = getAdapter(self.representative_object, IBuyableContent)
         raw_price = buyable.price
         member = self.pmt.getAuthenticatedMember()
-        return (self.cr.queryCredit(member.getId(), IOneWeekPublishedCredit.__identifier__) / Decimal('%0.2f' % raw_price))
+        return int(self.cr.queryCredit(member.getId(), IOneWeekPublishedCredit.__identifier__) / Decimal('%0.2f' % raw_price))
 
     def cashvalue(self):
         tax = getUtility(ITaxUtility)
